WIP Mobile menu screen pushover
Because:
- Menu should share screen with rest of content
- Better to slide in/out menu from side screen
- Nice to limit scrolling to veritcal scroll

The most important stuff done here was to implement a mobile Nav
that pushes over the rest of the content when it is open.

Also created a context for the open state of the mobile Nav since
this info will be shared among several other components in the App.